requirements.txt
flask-sqlalchemy==2.4.1
flask-celery==2.4.3
redis==3.4.1
beautifulsoup4==4.8.2
requests==2.23.0
fake-useragent==0.1.11
PyMySQL==0.9.3